requirements.txt
numpy
scikit-learn
dill
google-generativeai
